+++ API/Controllers/ProjectController.cs
@@ -15,6 +15,7 @@
 * If not, see https://www.gnu.org/licenses/lgpl-3.0.txt
 */
 
+using API.Extensions;
 using API.Resources;
 using AutoMapper;
 using Microsoft.AspNetCore.Authorization;
@@ -44,7 +45,7 @@ namespace API.Controllers
         private readonly IUserService userService;
 
         /// <summary>
-        ///     Initialize a new instance of ProjectController
+        /// Initialize a new instance of ProjectController
         /// </summary>
         /// <param name="projectService"></param>
         /// <param name="userService"></param>
@@ -117,7 +118,7 @@ namespace API.Controllers
         /// </summary>
         /// <returns></returns>
         [HttpPost]
-        [Authorize(Policy = nameof(Defaults.Scopes.ProjectWrite))]
+        [Authorize]
         public async Task<IActionResult> CreateProjectAsync([FromBody] ProjectResource projectResource)
         {
             if(projectResource == null)
@@ -132,10 +133,8 @@ namespace API.Controllers
             }
             Project project = mapper.Map<ProjectResource, Project>(projectResource);
 
-            //TODO: When login in frontend is functioning, get the user from JWT information
-            User user = await userService.GetUserAsync(1);
+            User user = await userService.GetUserAsync(projectResource.UserId);
             project.User = user;
-            project.UserId = user.Id;
             try
             {
                 projectService.Add(project);
@@ -160,7 +159,7 @@ namespace API.Controllers
         /// <param name="projectResource"></param>
         /// <returns></returns>
         [HttpPut("{projectId}")]
-        [Authorize(Policy = nameof(Defaults.Scopes.ProjectWrite))]
+        [Authorize]
         public async Task<IActionResult> UpdateProject(int projectId, [FromBody] ProjectResource projectResource)
         {
             Project project = await projectService.FindAsync(projectId);
@@ -177,9 +176,23 @@ namespace API.Controllers
 
             mapper.Map(projectResource, project);
 
+            string identity = HttpContext.User.GetStudentId(HttpContext);
+            bool isAllowed = userService.UserHasScope(identity, nameof(Defaults.Scopes.ProjectWrite));
+            User user = await userService.GetUserByIdentityIdAsync(identity);
+
+            if(!(project.UserId == user.Id || isAllowed))
+            {
+                ProblemDetails problem = new ProblemDetails
+                {
+                    Title = "Failed to edit the project.",
+                    Detail = "The user is not allowed to edit the project.",
+                    Instance = "2E765D18-8EBC-4117-8F9E-B800E8967038"
+                };
+                return Unauthorized(problem);
+            }
+
             projectService.Update(project);
             projectService.Save();
-
             return Ok(mapper.Map<Project, ProjectResourceResult>(project));
         }
 
@@ -188,10 +201,11 @@ namespace API.Controllers
         /// </summary>
         /// <returns></returns>
         [HttpDelete("{projectId}")]
-        [Authorize(Policy = nameof(Defaults.Scopes.ProjectWrite))]
+        [Authorize]
         public async Task<IActionResult> DeleteProject(int projectId)
         {
-            if(await projectService.FindAsync(projectId) == null)
+            Project project = await projectService.FindAsync(projectId);
+            if(project == null)
             {
                 ProblemDetails problem = new ProblemDetails
                 {
@@ -201,11 +215,25 @@ namespace API.Controllers
                 };
                 return NotFound(problem);
             }
+            
+            string identity = HttpContext.User.GetStudentId(HttpContext);
+            bool isAllowed = userService.UserHasScope(identity, nameof(Defaults.Scopes.ProjectWrite));
+            User user = await userService.GetUserByIdentityIdAsync(identity);
+
+            if(!(project.UserId == user.Id || isAllowed))
+            {
+                ProblemDetails problem = new ProblemDetails
+                {
+                    Title = "Failed to delete the project.",
+                    Detail = "The user is not allowed to delete the project.",
+                    Instance = "D0363680-5B4F-40A1-B381-0A7544C70164"
+                };
+                return Unauthorized(problem);
+            }
+
             await projectService.RemoveAsync(projectId);
             projectService.Save();
             return Ok();
         }
-
     }
-
 }