+++ IdentityServer/Startup.cs
@@ -79,7 +79,7 @@ namespace IdentityServer
                                                      .AddTestUsers(TestUsers.Users);
 
             // in-memory, code config
-            builder.AddInMemoryIdentityResources(IdentityConfig.Ids);
+            builder.AddInMemoryIdentityResources(IdentityConfig.GetIdentityResources());
             builder.AddInMemoryApiResources(IdentityConfig.Apis);
             builder.AddInMemoryClients(IdentityConfig.Clients(Config));
             builder.AddTestUsers(TestUsers.Users);
@@ -93,10 +93,12 @@ namespace IdentityServer
             //         // ...
             //     });
 
+            //TODO: Have some sort of certificate on the production servers
+            // not recommended for production - you need to store your key material somewhere secure
+            builder.AddDeveloperSigningCredential();
+
             if(Environment.IsDevelopment())
             {
-                //TODO: Have some sort of certificate on the production servers
-                // not recommended for production - you need to store your key material somewhere secure
                 builder.AddDeveloperSigningCredential();
             }
             services.AddCors(options =>